refactor(calc): パッケージ周りの修正
