+++ app/src/main/java/com/example/yosuke/calculator/di/modules/CalcModule.kt
@@ -1,7 +1,7 @@
 package com.example.yosuke.calculator.di.modules
 
 import android.arch.lifecycle.ViewModel
-import com.example.yosuke.calculator.CalcViewModel
+import com.example.yosuke.calculator.viewmodel.CalcViewModel
 import com.example.yosuke.calculator.di.ViewModelKey
 import dagger.Binds
 import dagger.Module
