+++ app/src/main/java/com/example/yosuke/calculator/di/modules/MainActivityModule.kt
@@ -1,6 +1,6 @@
 package com.example.yosuke.calculator.di.modules
 
-import com.example.yosuke.calculator.MainActivity
+import com.example.yosuke.calculator.view.MainActivity
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
