+++ app/src/main/res/layout/activity_main.xml
@@ -5,12 +5,12 @@
         <import type="android.text.TextUtils"/>
         <variable
                 name="viewModel"
-                type="com.example.yosuke.calculator.CalcViewModel"/>
+                type="com.example.yosuke.calculator.viewmodel.CalcViewModel"/>
     </data>
     <android.support.constraint.ConstraintLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context=".MainActivity">
+            tools:context=".view.MainActivity">
 
         <TextView
                 android:layout_width="wrap_content"
