+++ app/src/main/res/layout/fragment_calc_button.xml
@@ -4,7 +4,7 @@
     <data>
         <variable
                 name="viewModel"
-                type="com.example.yosuke.calculator.CalcViewModel"/>
+                type="com.example.yosuke.calculator.viewmodel.CalcViewModel"/>
     </data>
     <android.support.constraint.ConstraintLayout
             android:layout_width="match_parent"
